fix: 🐛 consolidate
