+++ src/components/ui/meeting-card.tsx
@@ -1,7 +1,5 @@
 import AccessTimeIcon from "@mui/icons-material/AccessTime";
 import DateRangeIcon from "@mui/icons-material/DateRange";
-import DeleteForever from "@mui/icons-material/DeleteForever";
-import ExitToApp from "@mui/icons-material/ExitToApp";
 import FmdGoodIcon from "@mui/icons-material/FmdGood";
 import GroupIcon from "@mui/icons-material/Group";
 import MoreVertIcon from "@mui/icons-material/MoreVert";
@@ -20,24 +18,13 @@ import {
 	Typography,
 } from "@mui/material";
 import Link from "next/link";
-import { type ElementType, useState } from "react";
-import { DeleteModal } from "@/components/availability/header/delete-modal";
-import type { SelectMeeting } from "@/db/schema";
+import { type ElementType, useId, useState } from "react";
+import type { MeetingCardViewModel } from "@/lib/meeting-card/mapper";
+import { getDeleteLeaveAction } from "@/lib/meetings/delete-leave-action";
 
-interface MeetingCardProps {
-	meeting: SelectMeeting;
-	meetingName: string;
-	meetingOrganizer: string;
-	dateStart: string;
-	dateEnd: string;
-	timeStart: string;
-	timeEnd: string;
-	numResponders: number;
-	location?: string | null;
-	scheduled?: boolean;
-	scheduledLabel?: string;
-	meetingLink: string;
+interface MeetingCardProps extends MeetingCardViewModel {
 	isOwner: boolean;
+	onDeleteLeave?: () => void;
 }
 
 const metaIconSx = { fontSize: 16, color: "text.secondary", flexShrink: 0 };
@@ -66,7 +53,6 @@ const MetaItem = ({
 );
 
 const MeetingCard = ({
-	meeting,
 	meetingName,
 	meetingOrganizer,
 	dateStart,
@@ -79,20 +65,22 @@ const MeetingCard = ({
 	scheduledLabel,
 	meetingLink,
 	isOwner,
+	onDeleteLeave,
 }: MeetingCardProps) => {
+	const menuId = useId();
+	const { label: actionLabel, Icon, menuColor } = getDeleteLeaveAction(isOwner);
+
 	const dateLabel =
 		dateStart && dateEnd && dateStart !== dateEnd
 			? `${dateStart} - ${dateEnd}`
 			: dateStart;
 
 	const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
-	const [isDeleteModalOpen, setIsDeleteModalOpen] = useState(false);
-	const [isDeletionPending, setIsDeletionPending] = useState(false);
 	const open = Boolean(anchorEl);
 
-	const handleOpenDeleteModal = () => {
+	const handleOpenDeleteLeave = () => {
 		setAnchorEl(null);
-		setIsDeleteModalOpen(true);
+		onDeleteLeave?.();
 	};
 
 	return (
@@ -130,54 +118,44 @@ const MeetingCard = ({
 							{meetingOrganizer}
 						</Typography>
 					</Box>
-					<IconButton
-						size="small"
-						edge="end"
-						sx={{ mt: -0.5, flexShrink: 0 }}
-						aria-label="Meeting options"
-						onClick={(e) => setAnchorEl(e.currentTarget)}
-					>
-						<MoreVertIcon sx={{ fontSize: { xs: 24, sm: 20 } }} />
-					</IconButton>
+					{onDeleteLeave && (
+						<>
+							<IconButton
+								size="small"
+								edge="end"
+								sx={{ mt: -0.5, flexShrink: 0 }}
+								aria-label="Meeting options"
+								aria-haspopup="true"
+								aria-expanded={open}
+								aria-controls={open ? menuId : undefined}
+								onClick={(e) => setAnchorEl(e.currentTarget)}
+							>
+								<MoreVertIcon sx={{ fontSize: { xs: 24, sm: 20 } }} />
+							</IconButton>
 
-					<Menu
-						anchorEl={anchorEl}
-						open={open}
-						onClose={() => setAnchorEl(null)}
-						anchorOrigin={{ vertical: "bottom", horizontal: "right" }}
-						transformOrigin={{ vertical: "top", horizontal: "right" }}
-						slotProps={{
-							paper: {
-								style: {
-									maxHeight: 48 * 4.5,
-									minWidth: 180,
-								},
-							},
-						}}
-					>
-						<MenuItem
-							onClick={handleOpenDeleteModal}
-							sx={{ color: isOwner ? "error.main" : "warning.main" }}
-						>
-							<ListItemIcon sx={{ color: "inherit", minWidth: 36 }}>
-								{isOwner ? (
-									<DeleteForever fontSize="small" />
-								) : (
-									<ExitToApp fontSize="small" />
-								)}
-							</ListItemIcon>
-							{isOwner ? "Delete Meeting" : "Leave Meeting"}
-						</MenuItem>
-					</Menu>
-
-					<DeleteModal
-						meetingData={meeting}
-						isOpen={isDeleteModalOpen}
-						handleOpenChange={setIsDeleteModalOpen}
-						isOwner={isOwner}
-						isDeletionPending={isDeletionPending}
-						onDeletionPendingChange={setIsDeletionPending}
-					/>
+							<Menu
+								id={menuId}
+								anchorEl={anchorEl}
+								open={open}
+								onClose={() => setAnchorEl(null)}
+								anchorOrigin={{ vertical: "bottom", horizontal: "right" }}
+								transformOrigin={{ vertical: "top", horizontal: "right" }}
+								slotProps={{
+									paper: { sx: { minWidth: 180 } },
+								}}
+							>
+								<MenuItem
+									onClick={handleOpenDeleteLeave}
+									sx={{ color: menuColor }}
+								>
+									<ListItemIcon sx={{ color: "inherit", minWidth: 36 }}>
+										<Icon fontSize="small" />
+									</ListItemIcon>
+									{actionLabel}
+								</MenuItem>
+							</Menu>
+						</>
+					)}
 				</Box>
 
 				{scheduled && scheduledLabel ? (
